chore : requestparam, requestpart, pathvariable 등 없을 경우 에러 처리

+++ src/main/java/project/backend/domain/notice/controller/NoticeController.java
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.*;
 import project.backend.domain.notice.dto.NoticePostRequestDto;
 import project.backend.domain.notice.dto.NoticeResponseDto;
@@ -13,6 +14,8 @@ import project.backend.domain.notice.dto.NoticePatchRequestDto;
 import project.backend.domain.notice.entity.Notice;
 import project.backend.domain.notice.mapper.NoticeMapper;
 import project.backend.domain.notice.service.NoticeService;
+import project.backend.global.error.exception.BusinessException;
+import project.backend.global.error.exception.ErrorCode;
 
 import java.util.List;
 
@@ -27,14 +30,20 @@ public class NoticeController {
 
 
     @PostMapping
-    public ResponseEntity postNotice(@RequestBody NoticePostRequestDto noticePostRequestDto) {
+    public ResponseEntity postNotice(@RequestBody(required = false) NoticePostRequestDto noticePostRequestDto) {
+        if (ObjectUtils.isEmpty(noticePostRequestDto)){
+            throw new BusinessException(ErrorCode.MISSING_REQUEST);
+        }
         Notice notice = noticeService.createNotice(noticePostRequestDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(noticeMapper.noticeToNoticeResponseDto(notice));
     }
 
     @ApiOperation(value = "공지 목록")
     @GetMapping("/{noticeId}")
-    public ResponseEntity getNotice(@PathVariable Long noticeId) {
+    public ResponseEntity getNotice(@PathVariable(required = false) Long noticeId) {
+        if (ObjectUtils.isEmpty(noticeId)){
+            throw new BusinessException(ErrorCode.MISSING_REQUEST);
+        }
         NoticeResponseDto noticeResponseDto = noticeMapper.noticeToNoticeResponseDto(noticeService.getNotice(noticeId));
         return ResponseEntity.status(HttpStatus.OK).body(noticeResponseDto);
     }
@@ -47,14 +56,20 @@ public class NoticeController {
 
     @PatchMapping("/{noticeId}")
     public ResponseEntity putNotice(
-            @PathVariable Long noticeId,
-            @RequestBody NoticePatchRequestDto noticePatchRequestDto) {
+            @PathVariable(required = false) Long noticeId,
+            @RequestBody(required = false) NoticePatchRequestDto noticePatchRequestDto) {
+        if (ObjectUtils.isEmpty(noticeId) || ObjectUtils.isEmpty(noticePatchRequestDto)){
+            throw new BusinessException(ErrorCode.MISSING_REQUEST);
+        }
         NoticeResponseDto noticeResponseDto = noticeMapper.noticeToNoticeResponseDto(noticeService.patchNotice(noticeId, noticePatchRequestDto));
         return ResponseEntity.status(HttpStatus.OK).body(noticeResponseDto);
     }
 
     @DeleteMapping("/{noticeId}")
-    public ResponseEntity deleteNotice(@PathVariable Long noticeId) {
+    public ResponseEntity deleteNotice(@PathVariable(required = false) Long noticeId) {
+        if (ObjectUtils.isEmpty(noticeId)){
+            throw new BusinessException(ErrorCode.MISSING_REQUEST);
+        }
         noticeService.deleteNotice(noticeId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
     }
